Added features to makesymlinks.sh
Added more files to list of files to copy; adjusted script to not
attempt to copy non-existent files.

+++ makesymlinks.sh
@@ -8,7 +8,8 @@
 
 dir=~/dotfiles                    # dotfiles directory
 olddir=~/dotfiles_old$$             # old dotfiles backup directory
-files="bashrc vimrc vim zshrc oh-my-zsh"    # list of files/folders to symlink in homedir
+# list of files/folders to symlink in homedir
+files="bashrc bash_alias bash_profile kshrc profile vimrc vim zshrc oh-my-zsh "
 
 ##########
 
@@ -24,8 +25,10 @@ echo "...done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks 
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file $olddir
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/.$file
+    if [ -f $file ] ; then
+        echo "Moving previous version of $file from ~ to $olddir"
+        mv ~/.$file $olddir
+        echo "Creating symlink to $file in home directory."
+        ln -s $dir/$file ~/.$file
+    fi
 done
